Update using-sentry.mdx
After running into build issues upgrading to react-native-sentry from expo-sentry alongside my SDK 51 upgrade, I noticed that in Sentry’s documentation the env variable SENTRY_PROJECT which this documentation proposes as an alternative to the project key in the sentry plugin’s configuration, maps to the sentry project’s slug, not the name. After rebuilding my project with eas after adjusting the value to use the slug, it works. Perhaps most developers' project name & slug are the same so I’m among the few or only to run into this yet?